refactor: 테스트 메서드 리팩토링
